requirements.txt
google-cloud-storage>=1.16.1
google-cloud-kms>=1.1.0
